% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrImpute.R
\name{dist2}
\alias{dist2}
\title{dist2}
\usage{
dist2(X, method)
}
\arguments{
\item{X}{Log transformed gene expression matrix (Gene by Cell).}

\item{Distance}{method}
}
\value{
A matrix object
}
\description{
Computing the distance matrix
}
\references{
Wuming Gong, Il-Youp Kwak, Pruthvi Pota, Kaoko Koyano-Nakagawa and Daniel J. Garry (2017+)
DrImpute: Imputing dropout eveents in single cell RNA sequencing data
}
\author{
Wuming Gong
}
