% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrImpute.R
\name{DrImpute}
\alias{DrImpute}
\title{DrImpute}
\usage{
DrImpute(X, ks = 10:15, dists = c("spearman", "pearson"), fast = FALSE,
  dropout.probability.threshold = 0, n.dropout = 10000,
  n.background = 10000, mc.cores = 1)
}
\arguments{
\item{X}{Gene expression matrix (gene by cell).}

\item{ks}{Number of cell clustering groups. Default set to ks = 10:15.}

\item{dists}{Distribution matrices to use. Default is set to c("spearman", "pearson"). "eucleadian" can be added as well.}

\item{fast}{Whether or not using approximation for large scale scRNA-seq data (default: FALSE)}

\item{dropout.probability.threshold}{This argument determine the dropout events for imputation.  If this argument is zero, all
zero events will be considered for imputation; otherwise, only the zero entries that have greater probability than the specified threshold
will be imputed (default: 0)}

\item{n.dropout}{Number of dropout entries for evaluting the dropout probability (default: 10000)}

\item{n.background}{Number of background entries for evaluting the dropout probability (default: 10000)}

\item{mc.cores}{Number of CPU cores for the fast method (default: 1)}
}
\value{
A matrix object
}
\description{
Imputing dropout events in single-cell RNA-sequencing data.
}
\examples{

library(scDatasets)
library(SummarizedExperiment)
data(usoskin)
X <- assays(usoskin)$count
X <- preprocess(X, min.expressed.gene = 0)
X.log <- log(X + 1)
set.seed(1)
X.imp <- DrImpute(X.log)

}
\references{
Wuming Gong, Il-Youp Kwak, Pruthvi Pota, Kaoko Koyano-Nakagawa and Daniel J. Garry (2017+)
DrImpute: Imputing dropout eveents in single cell RNA sequencing data
}
\author{
Wuming Gong and Il-Youp Kwak
}
