% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrImpute.R
\name{estimate.expression}
\alias{estimate.expression}
\title{estimate.expression}
\usage{
estimate.expression(X, ks = 10:15, dists = c("spearman", "pearson"),
  dim.reduc.prop = 0.05, max.dim = 100, pc.cdr.cc.cutoff = 1,
  mc.cores = 1, batch.size = NA)
}
\arguments{
\item{X}{Log transformed gene expression matrix (Gene by Cell).}

\item{ks}{Number of cell clustering groups. (default: 10:15)}

\item{dists}{Distribution matrices to use. (default: c("spearman", "pearson"))}

\item{dim.reduc.prop}{Proportion of principal components to use for K-means clustering. (default: 0.05)}

\item{max.dim}{Maximum dimensions for PCA (default: 100)}

\item{pc.cdr.cc.cutoff}{The PC's which correlation coefficient with CDR (cellular detection rate) will be removed. (default: 1)}

\item{mc.cores}{Number of CPU cores (default: 1)}

\item{batch.size}{Batch size of sampling based MDS approximation}
}
\value{
A matrix object
}
\description{
Estimate the expected expression levels for entries in the gene by cell expression matrix
}
\references{
Wuming Gong, Il-Youp Kwak, Pruthvi Pota, Kaoko Koyano-Nakagawa and Daniel J. Garry (2017+)
DrImpute: Imputing dropout eveents in single cell RNA sequencing data
}
\author{
Wuming Gong
}
