% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{imp0clC}
\alias{imp0clC}
\alias{imp0clC,DrImpute-method}
\title{imp0clC c++ implementation to impute the zeros}
\usage{
imp0clC(X, cls)
}
\arguments{
\item{X}{the matrix of data}

\item{cls}{the grouping information}
}
\value{
a matrix of expression values
}
\description{
impute the zeros in a single cell dataset based on a matrix and a grouping information
}
