% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrImpute.R
\name{kmeans2}
\alias{kmeans2}
\title{kmeans2}
\usage{
kmeans2(x, centers, batch.size = 10000, iter.max = 1e+09)
}
\arguments{
\item{x}{An input matrix}

\item{centers}{Number of centers}

\item{batch.size}{Batch size}

\item{iter.max}{Maximum iterations}
}
\description{
Mini-batch k-means
}
\references{
Il-Youp Kwak, Wuming Gong, Kaoko Koyano-Nakagawa and Daniel J. Garry (2017+)
DrImpute: Imputing dropout eveents in single cell RNA sequencing data

D. Schully (2010) 
Web-scale k-means clustering
}
