% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrImpute.R
\name{pca.dist.matrix}
\alias{pca.dist.matrix}
\title{pca.dist.matrix}
\usage{
pca.dist.matrix(X, k, dist.method = "spearman", mc.cores = 1,
  batch.size = 1000)
}
\arguments{
\item{X}{Log transformed gene expression matrix (Gene by Cell).}

\item{k}{Number of PCs}

\item{dist.method}{Distance method (default: spearman)}

\item{mc.cores}{Number of CPU cores (default: 1)}

\item{batch.size}{Batch size of sampling based MDS approximation}
}
\value{
A matrix object
}
\description{
PCA of distance matrix
}
\references{
Wuming Gong, Il-Youp Kwak, Pruthvi Pota, Kaoko Koyano-Nakagawa and Daniel J. Garry (2017+)
DrImpute: Imputing dropout eveents in single cell RNA sequencing data
}
\author{
Wuming Gong
}
