% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DrImpute.R
\name{prcomp2}
\alias{prcomp2}
\title{prcomp2}
\usage{
prcomp2(X, n = NA, ...)
}
\arguments{
\item{X}{Log transformed gene expression matrix (Gene by Cell).}

\item{n}{Number of PCs}
}
\value{

}
\description{
A wrapper for PCA
}
\references{
Wuming Gong, Il-Youp Kwak, Pruthvi Pota, Kaoko Koyano-Nakagawa and Daniel J. Garry (2017+)
DrImpute: Imputing dropout eveents in single cell RNA sequencing data
}
\author{
Wuming Gong
}
